Break out extractor to separate package.
